+++ lib/features/home/sub_screens/profiles/widgets/profile_card.dart
@@ -11,6 +11,7 @@ import 'package:ebalistyka/shared/widgets/text_input_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
+import 'package:ebalistyka/shared/widgets/dividers.dart';
 
 class ProfileCard extends ConsumerStatefulWidget {
   const ProfileCard({
@@ -197,7 +198,7 @@ class _ProfileCardState extends ConsumerState<ProfileCard> {
 
     return Column(
       children: [
-        const Divider(height: 1),
+        const TileDivider(),
         ListSectionTile(
           "Ammo",
           onTap: widget.onEditAmmo,
@@ -245,7 +246,7 @@ class _ProfileCardState extends ConsumerState<ProfileCard> {
 
     return Column(
       children: [
-        const Divider(height: 1),
+        const TileDivider(),
         ListSectionTile(
           "Sight",
           onTap: widget.onEditSight,